requirements.txt
discord.py>=2.3.0
httpx[http2]>=0.26.0
PyYAML>=6.0
